Files restructure and names changed for english
