--- README.md
@@ -1,33 +1,0 @@
-# bedroom-design
-This is an assignment done for my Computer Graphics lab using C++ and OpenGL. The GL, GLU and GLUT libraries were used. The bedroom was designed using cubes, pyramids and one 2D polygon.
-
-![picture alt](https://github.com/n-gauhar/bedroom-design/blob/main/Sample%20image.png)
-
-The features of the bedroom - 
-* No lighting, no textures, no curve
-* A pendulum clock with a moving pendulum
-* A rotating fan (can be turned on/off)
-* Change eyepoint and lookat point using keys
-
-#### The keys used are -  
-
-To move Eye point:
-1. w: up
-2. s: down
-3. a: left
-4. d: right
-5. i: zoom in
-6. o: zoom out
-   
-To move Camera point:
-1. j: up
-2. n: down
-3. b: left
-4. m: right
-5. l: move nearer
-6. k: move far
-    
-
-1. q to move to default position (both eyepoint & lookat point)
-2. f: to turn on/off fan
-
