+++ includes/simpleDraw.h
@@ -1,0 +1,9 @@
+#ifndef SIMPLE_DRAW_H
+#define SIMPLE_DRAW_H
+
+void drawCube();
+void drawPyramid();
+void polygon();
+void polygonLine();
+
+#endif
