+++ src/simpleDraw.cpp
@@ -1,0 +1,122 @@
+#include<simpleDraw.h>
+#include <GL/glew.h>
+#include <GL/glut.h>
+
+static GLfloat v_cube[8][3] = {
+    {0.0, 0.0, 0.0}, //0
+    {0.0, 0.0, 3.0}, //1
+    {3.0, 0.0, 3.0}, //2
+    {3.0, 0.0, 0.0}, //3
+    {0.0, 3.0, 0.0}, //4
+    {0.0, 3.0, 3.0}, //5
+    {3.0, 3.0, 3.0}, //6
+    {3.0, 3.0, 0.0}  //7
+};
+
+static GLubyte quadIndices[6][4] = {
+    {0, 1, 2, 3}, //bottom
+    {4, 5, 6, 7}, //top
+    {5, 1, 2, 6}, //front
+    {0, 4, 7, 3}, // back is clockwise
+    {2, 3, 7, 6}, //right
+    {1, 5, 4, 0}  //left is clockwise
+};
+
+/*
+static GLfloat colors[6][3] = {
+    {0.4, 0.1, 0.0}, //bottom
+    {0.6, 0.0, 0.7}, //top
+    {0.0, 1.0, 0.0},
+    {0.0, 1.0, 1.0},
+    {0.8, 0.0, 0.0},
+    {0.3, 0.6, 0.7}
+};
+*/
+static GLfloat v_pyramid[5][3] = {
+    {0.0, 0.0, 0.0},
+    {0.0, 0.0, 2.0},
+    {2.0, 0.0, 2.0},
+    {2.0, 0.0, 0.0},
+    {1.0, 4.0, 1.0}
+};
+
+static GLubyte p_Indices[4][3] = {
+    {4, 1, 2},
+    {4, 2, 3},
+    {4, 3, 0},
+    {4, 0, 1}
+};
+
+static GLubyte PquadIndices[1][4] = {
+    {0, 3, 2, 1}
+};
+
+
+void drawCube() {
+    glBegin(GL_QUADS);
+    for (GLint i = 0; i <6; i++)
+    {
+        glVertex3fv(&v_cube[quadIndices[i][0]][0]);
+        glVertex3fv(&v_cube[quadIndices[i][1]][0]);
+        glVertex3fv(&v_cube[quadIndices[i][2]][0]);
+        glVertex3fv(&v_cube[quadIndices[i][3]][0]);
+    }
+    glEnd();
+}
+
+void drawPyramid(){
+    glBegin(GL_TRIANGLES);
+    for (GLint i = 0; i <4; i++)
+    {
+        glVertex3fv(&v_pyramid[p_Indices[i][0]][0]);
+        glVertex3fv(&v_pyramid[p_Indices[i][1]][0]);
+        glVertex3fv(&v_pyramid[p_Indices[i][2]][0]);
+    }
+    glEnd();
+
+    glBegin(GL_QUADS);
+    for (GLint i = 0; i <1; i++)
+    {
+        glVertex3fv(&v_pyramid[PquadIndices[i][0]][0]);
+        glVertex3fv(&v_pyramid[PquadIndices[i][1]][0]);
+        glVertex3fv(&v_pyramid[PquadIndices[i][2]][0]);
+        glVertex3fv(&v_pyramid[PquadIndices[i][3]][0]);
+    }
+    glEnd();
+
+}
+
+void polygon(){
+    glBegin(GL_POLYGON);
+    glVertex2f(0,0);
+    glVertex2f(6,0);
+    glVertex2f(5.8,1);
+    glVertex2f(5.2,2);
+    glVertex2f(5, 2.2);
+    glVertex2f(4, 2.8);
+    glVertex2f(3,3);
+    glVertex2f(2, 2.8);
+    glVertex2f(1, 2.2);
+    glVertex2f(0.8, 2);
+    glVertex2f(0.2,1);
+    
+    glEnd();
+}
+
+void polygonLine(){
+    glBegin(GL_LINE_STRIP);
+    //glVertex2f(0,0);
+    glVertex2f(6,0);
+    glVertex2f(5.8,1);
+    glVertex2f(5.2,2);
+    glVertex2f(5, 2.2);
+    glVertex2f(4, 2.8);
+    glVertex2f(3,3);
+    glVertex2f(2, 2.8);
+    glVertex2f(1, 2.2);
+    glVertex2f(0.8, 2);
+    glVertex2f(0.2,1);
+    glVertex2f(0,0);
+    
+    glEnd();
+}
